Refactor icon color logic to separate utility file
Moved the getColorBasedOnValue function from StopsPage.jsx to a new utils/mapIcons.js file for better modularity and reuse. Updated imports and adjusted related code accordingly.
